separate type file for TabPanel
