+++ src/types/tabPanel.type.ts
@@ -1,0 +1,11 @@
+type TabItem = {
+	label: string;
+	value: string;
+}
+
+export interface TabPanelProps {
+	tabs: TabItem[];
+	value: string;
+	onChange: (value: string) => void;
+	className?: string;
+}
